test1

+++ app/models/cliente.model.js
@@ -13,9 +13,8 @@ module.exports = (sequelize, Sequelize) => {
         apellidos: {
             type: Sequelize.STRING(20)
         },
-        // En cliente.model.js
         dpi: {
-            type: Sequelize.BIGINT,  // Cambiado de INTEGER a BIGINT
+            type: Sequelize.INTEGER,
             unique: true,
             allowNull: false
         },
